splitted joincoffee and joinsomeone tests
