+++ Test/JoinCoffeeCommandTest.cs
@@ -62,62 +62,5 @@ namespace Test
 
             _slack.Verify(slack => slack.SendMessage(It.IsAny<string>(), Channel));
         }
-
-        [TestMethod]
-        public async Task GivenUserIsAdmin_WhenJoiningOtherInCoffee_UserIsAddedToRoster()
-        {
-            await _coffeeCommands.JoinSomeone(command);
-
-            _channelRepository.Verify(channelRepo => channelRepo.AddToRoster("otherUserid"));
-        }
-
-        [TestMethod]
-        public async Task GivenUserIsAdmin_WhenJoiningOtherInCoffee_UserIsAddedToDatabase()
-        {
-            await _coffeeCommands.JoinSomeone(command);
-
-            _userRepository.Verify(userRepo => userRepo.AddUser("otherUserid"));
-        }
-
-        [TestMethod]
-        public async Task GivenUserIsAdmin_WhenJoiningOtherInCoffee_BroadcastIsSentToChannel()
-        {
-            await _coffeeCommands.JoinSomeone(command);
-
-            _slack.Verify(slack => slack.SendMessage(It.IsAny<string>(), Channel));
-        }
-
-        [TestMethod]
-        [ExpectedException(typeof(UserNotAdminException))]
-        public async Task GivenUserIsNotAdmin_WhenJoiningOtherInCoffee_DontAddUserToRoster()
-        {
-            _adminValidator.Setup(validator => validator.ValidateUserIsAdmin(User)).Throws(new UserNotAdminException());
-
-            await _coffeeCommands.JoinSomeone(command);
-
-            _channelRepository.Verify(channelRepo => channelRepo.AddToRoster(It.IsAny<string>()), Times.Never());
-        }
-
-        [TestMethod]
-        [ExpectedException(typeof(UserNotAdminException))]
-        public async Task GivenUserIsNotAdmin_WhenJoiningOtherInCoffee_DontAddUser()
-        {
-            _adminValidator.Setup(validator => validator.ValidateUserIsAdmin(User)).Throws(new UserNotAdminException());
-
-            await _coffeeCommands.JoinSomeone(command);
-
-            _userRepository.Verify(userRepo => userRepo.AddUser(It.IsAny<string>()), Times.Never());
-        }
-
-        [TestMethod]
-        [ExpectedException(typeof(UserNotAdminException))]
-        public async Task GivenUserIsNotAdmin_WhenJoiningOtherInCoffee_DontBroadcastMessage()
-        {
-            _adminValidator.Setup(validator => validator.ValidateUserIsAdmin(User)).Throws(new UserNotAdminException());
-
-            await _coffeeCommands.JoinSomeone(command);
-
-            _slack.Verify(slack => slack.SendMessage(It.IsAny<string>(), Channel), Times.Never());
-        }
     }
 }
